Add support to read cta dl1 files
- read_telescope_data now distinguisches based on the data_format key
- support for the intermediate multi_telescope format is removed with
this
- chunkwise reading works, but small chunks are somewhat inefficient
if the pointing, focal length, mc columns need to be loaded as well
because these are in different tables and need to be matched
- all columns of the dataframes get loaded initially because
not all wanted columns are necessarily in the /.../parameters tables.
- chunks and sampling is relative to telescope events not array events
- predictions made on cta data can be dropped with the
drop_prediction_groups method. This is currently very hardcoded.
Adding predictions is done in the apply scripts for now, this
could be put into a new function (or replace the append_column_to_hdf5
function with smth more generic)